Parse Instacart delivery info page
This gives us more reliable delivery info for Instacart, as the site
seems to sometimes show "See delivery times" in the storefront page
regardless of whether slots are available.

(Seems like they have rolled out tweaks so the <div> I'm targeting is
more scrapeable, but we can't count on that LOL.)

Switch WebDriver to RemoteWebDriver so that we have the interface to
execute JS (for innerHTML shenanigans).